See README for upgrades/fixes

+++ lens.cpp
@@ -7866,12 +7866,12 @@ void Lens::multinest(const bool resume_sampling)
 #endif
 }
 
-void Lens::polychord()
+void Lens::polychord(const bool resume_previous)
 {
 #ifdef USE_POLYCHORD
 	if (setup_fit_parameters(true)==false) return;
 	fit_set_optimizations();
-	if ((mpi_id==0) and (fit_output_dir != ".")) {
+	if ((mpi_id==0) and (!resume_previous) and (fit_output_dir != ".")) {
 		string rmstring = "if [ -e " + fit_output_dir + " ]; then rm -r " + fit_output_dir + "; fi";
 		if (system(rmstring.c_str()) != 0) warn("could not delete old output directory for nested sampling results"); // delete the old output directory and remake it, just in case there is old data that might get mixed up when running mkdist
 		// I should probably give the nested sampling output a unique extension like ".nest" or something, so that mkdist can't ever confuse it with twalk output in the same dir
@@ -7955,8 +7955,8 @@ void Lens::polychord()
 	settings.base_dir      = fit_output_dir.c_str();
 	settings.file_root     = fit_output_filename.c_str();
 
-	settings.write_resume  = false;
-	settings.read_resume   = false;
+	settings.write_resume  = true;
+	settings.read_resume   = resume_previous;
 	settings.write_live    = true;
 	settings.write_dead    = true;
 	settings.write_stats   = true;
